backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pydantic==1.10.15
pydantic[email]==1.10.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pillow==10.3.0
python-pptx==0.6.23
google-auth==2.35.0
